+++ changelog.md
@@ -8,6 +8,7 @@
 * Skeleton
   * Flee Distance changed (12 -> 16 blocks; 7 -> 8 blocks faster flee) and now configurable
 * Zombie Pearlers no longer need to be on ground to pearl
+* Improved Digger Zombies, they now keep mining multiple blocks before stopping, and they should get stuck less
 
 ## 1.2.4
 * Port to 1.18.2, requires InsaneLib 1.4.6